whisper_pipeline/requirements.txt
# Core Dependencies
openai-whisper
torch
torchaudio
soundfile
numpy
python-dotenv

# LLM (llama3.2)
ollama

# CSM Dependencies
transformers
accelerate
sentencepiece
protobuf
huggingface-hub

# Audio Capture (for live input)
sounddevice

# Optional utilities
tqdm
